This is a squashed set of four commits.  The original commit messages are below in commit order:
Add in a Rubocop configuration tuned from my usual and updated to work with latest Rubocop.  115 offenses found.

Fixed stylistic and layout changes from Rubocop, as well as a few syntax issues which were in likely rarely triggered code paths.  115 -> 5 offenses in Rubocop.

Add in dependencies for Rubocop, RSpec, and SimpleCov, add in a spec_helper, and add in correct tasks in Rakefile

Add in configuration for RSpec and CodeClimate.  You can add CodeClimate to public repos for free.

This closes Lyle-Tafoya/Steam-Categorizer#18